Moving forward to the latest HF transformers

+++ collections/nemo_nlp/nemo_nlp/data/datasets/joint_intent_slot.py
@@ -214,7 +214,7 @@ class BertJointIntentSlotDataset(Dataset):
     def __getitem__(self, idx):
         return (np.array(self.all_input_ids[idx]),
                 np.array(self.all_segment_ids[idx]),
-                np.array(self.all_input_mask[idx]),
+                np.array(self.all_input_mask[idx], dtype=np.long),
                 np.array(self.all_loss_mask[idx]),
                 np.array(self.all_subtokens_mask[idx]),
                 self.all_intents[idx],
@@ -263,6 +263,6 @@ class BertJointIntentSlotInferDataset(Dataset):
     def __getitem__(self, idx):
         return (np.array(self.all_input_ids[idx]),
                 np.array(self.all_segment_ids[idx]),
-                np.array(self.all_input_mask[idx], dtype=np.float32),
+                np.array(self.all_input_mask[idx], dtype=np.long),
                 np.array(self.all_loss_mask[idx]),
                 np.array(self.all_subtokens_mask[idx]))
